# server/openapi.yaml
openapi: 3.0.0
info:
  title: Parcelec API
  version: 1.0.0
paths:
  /session/{sessionID}/user/{userID}/bid:
    post:
      summary: post a new bid
      parameters:
        - in: path
          name: sessionID
          schema:
            type: string
          required: true
          description: Session UUID
        - in: path
          name: userID
          schema:
            type: string
          required: true
          description: User UUID
      requestBody:
        description: Bid to post
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - bid
              properties:
                bid:
                  $ref: '#/components/schemas/NewBidBody'
      responses:
        '201':
          description: bid inserted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BidResponse'
...
components:
  schemas:
    NewBidBody:
      type: object
      required:
        - type
        - volume
        - price
      properties:
        type:
          type: string
          enum:
            - sell
            - buy
        volume:
          type: integer
        price:
          type: integer
    BidResponse:
      type: object
      required:
        - id
      properties:
        id:
          type: string
...